requirements.txt
bcrypt>=4.2.0
deepface
pillow>=11.0.0